Chore: [BADA-227] types위치 통합 및 주석 설명 추가 #142
